# services/gateway/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: OlimpicApp — Tournament Platform API
  description: |
    Multi-sport tournament management platform.
    All endpoints require JWT authentication via httpOnly cookies unless marked as public.
    The gateway proxies requests to internal microservices.
  version: 1.0.0
  contact:
    name: Tournament Platform Team
servers:
  - url: http://localhost:3000
    description: Local development gateway
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
  schemas:
...
    Payment:
      type: object
      properties:
        id: { type: string, format: uuid }
        tournamentId: { type: string, format: uuid }
        teamId: { type: string, format: uuid }
        amount: { type: number, format: decimal }
        currency: { type: string }
        paymentMethod: { type: string, nullable: true }
        reference: { type: string, nullable: true }
        status: { type: string, enum: [pending, confirmed, rejected, refunded] }
        paidAt: { type: string, format: date-time }
...
security:
  - cookieAuth: []
paths:
...
  /api/matches:
    get:
      tags: [Matches]
      summary: List matches
      parameters:
        - name: phaseId
          in: query
          schema: { type: string, format: uuid }
        - name: teamId
          in: query
          schema: { type: string, format: uuid }
        - name: status
          in: query
          schema: { type: string }
      responses:
        '200': { description: Match list }
    post:
      tags: [Matches]
      summary: Schedule a match
      responses:
        '201': { description: Match created }
  /api/matches/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Matches]
      summary: Get match by ID (includes periods)
      responses:
        '200': { description: Match details }

  # ── Standings ────────────────────────────────────
...
  /api/payments:
    get:
      tags: [Payments]
      summary: List payments
      parameters:
        - name: tournamentId
          in: query
          schema: { type: string, format: uuid }
        - name: teamId
          in: query
          schema: { type: string, format: uuid }
        - name: status
          in: query
          schema: { type: string, enum: [pending, confirmed, rejected, refunded] }
      responses:
        '200':
          description: Payment list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Payment' }
    post:
      tags: [Payments]
      summary: Record a payment (organizer+)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tournamentId, teamId, amount]
              properties:
                tournamentId: { type: string, format: uuid }
                teamId: { type: string, format: uuid }
                amount: { type: number, minimum: 0.01 }
                currency: { type: string, default: COP }
                paymentMethod: { type: string }
                reference: { type: string }
                notes: { type: string }
      responses:
        '201': { description: Payment recorded }
...